Fix IgnoreExampleErrors option to installPackage
We were checking in several places if there were example errors, and
only avoided raising an error when IgnoreExampleErrors was true in one
of these places.  So when an example had an error, we ended up raising
an error the next time it was checked.

We fix this by removing the example error check from the
"checkForErrors" helper function and move this code to right after we
generate the examples, when we were checking IgnoreExampleErrors.